+++ include/viua/bytecode/maps.h
@@ -35,11 +35,11 @@ const std::set<std::string> OP_MNEMONICS = {
     "nop",
 
     "izero",
-    "istore",
+    "integer",
     "iinc",
     "idec",
 
-    "fstore",
+    "float",
 
     "itof",
     "ftoi",
@@ -56,7 +56,7 @@ const std::set<std::string> OP_MNEMONICS = {
     "gte",
     "eq",
 
-    "strstore",
+    "string",
     "streq",
 
     "text",
@@ -68,7 +68,7 @@ const std::set<std::string> OP_MNEMONICS = {
     "textcommonsuffix",
     "textconcat",
 
-    "vec",
+    "vector",
     "vinsert",
     "vpush",
     "vpop",
@@ -167,11 +167,11 @@ const std::map<enum OPCODE, std::string> OP_NAMES = {
     {NOP, "nop"},
 
     {IZERO, "izero"},
-    {ISTORE, "istore"},
+    {INTEGER, "integer"},
     {IINC, "iinc"},
     {IDEC, "idec"},
 
-    {FSTORE, "fstore"},
+    {FLOAT, "float"},
 
     {ITOF, "itof"},
     {FTOI, "ftoi"},
@@ -188,7 +188,7 @@ const std::map<enum OPCODE, std::string> OP_NAMES = {
     {GTE, "gte"},
     {EQ, "eq"},
 
-    {STRSTORE, "strstore"},
+    {STRING, "string"},
     {STREQ, "streq"},
 
     {TEXT, "text"},
@@ -200,7 +200,7 @@ const std::map<enum OPCODE, std::string> OP_NAMES = {
     {TEXTCOMMONSUFFIX, "textcommonsuffix"},
     {TEXTCONCAT, "textconcat"},
 
-    {VEC, "vec"},
+    {VECTOR, "vector"},
     {VINSERT, "vinsert"},
     {VPUSH, "vpush"},
     {VPOP, "vpop"},