+++ include/viua/cg/bytecode/instructions.h
@@ -67,11 +67,11 @@ namespace cg {
         viua::internals::types::byte* opnop(viua::internals::types::byte*);
 
         viua::internals::types::byte* opizero(viua::internals::types::byte*, int_op);
-        viua::internals::types::byte* opistore(viua::internals::types::byte*, int_op, int_op);
+        viua::internals::types::byte* opinteger(viua::internals::types::byte*, int_op, int_op);
         viua::internals::types::byte* opiinc(viua::internals::types::byte*, int_op);
         viua::internals::types::byte* opidec(viua::internals::types::byte*, int_op);
 
-        viua::internals::types::byte* opfstore(viua::internals::types::byte*, int_op,
+        viua::internals::types::byte* opfloat(viua::internals::types::byte*, int_op,
                                                viua::internals::types::plain_float);
 
         viua::internals::types::byte* opitof(viua::internals::types::byte*, int_op, int_op);
@@ -89,7 +89,7 @@ namespace cg {
         viua::internals::types::byte* opgte(viua::internals::types::byte*, int_op, int_op, int_op);
         viua::internals::types::byte* opeq(viua::internals::types::byte*, int_op, int_op, int_op);
 
-        viua::internals::types::byte* opstrstore(viua::internals::types::byte*, int_op, std::string);
+        viua::internals::types::byte* opstring(viua::internals::types::byte*, int_op, std::string);
 
         viua::internals::types::byte* optext(viua::internals::types::byte*, int_op, std::string);
         viua::internals::types::byte* optext(viua::internals::types::byte*, int_op, int_op);
@@ -104,7 +104,7 @@ namespace cg {
                                                          int_op);
         viua::internals::types::byte* optextconcat(viua::internals::types::byte*, int_op, int_op, int_op);
 
-        viua::internals::types::byte* opvec(viua::internals::types::byte*, int_op, int_op, int_op);
+        viua::internals::types::byte* opvector(viua::internals::types::byte*, int_op, int_op, int_op);
         viua::internals::types::byte* opvinsert(viua::internals::types::byte*, int_op, int_op, int_op);
         viua::internals::types::byte* opvpush(viua::internals::types::byte*, int_op, int_op);
         viua::internals::types::byte* opvpop(viua::internals::types::byte*, int_op, int_op, int_op);