+++ src/cg/disassembler/disassembler.cpp
@@ -314,7 +314,7 @@ tuple<string, viua::internals::types::bytecode_size> disassembler::instruction(
     ostringstream oss;
     oss << opname;
 
-    if (op == STRSTORE) {
+    if (op == STRING) {
         ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
 
         ++ptr;  // for operand type
@@ -435,7 +435,7 @@ tuple<string, viua::internals::types::bytecode_size> disassembler::instruction(
         case BOOL:
             ptr = disassemble_ri_operand(oss, ptr);
             break;
-        case ISTORE:
+        case INTEGER:
             ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
             ptr = disassemble_ri_operand(oss, ptr);
 
@@ -526,7 +526,7 @@ tuple<string, viua::internals::types::bytecode_size> disassembler::instruction(
             }
 
             break;
-        case VEC:
+        case VECTOR:
         case VAT:
             ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
             ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
@@ -598,7 +598,7 @@ tuple<string, viua::internals::types::bytecode_size> disassembler::instruction(
             oss << dec;
 
             break;
-        case FSTORE:
+        case FLOAT:
             ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
 
             oss << ' ';